+++ src/components/core/DatePicker.js
@@ -2,9 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { withStyles } from '@material-ui/core/styles';
 import TextField from '@material-ui/core/TextField';
-import { timeFormat } from 'd3-time-format';
-
-const formatTime = timeFormat('%Y-%m-%d');
+import { formatDate } from '../../util/time';
 
 const styles = {
     root: {
@@ -12,15 +10,13 @@ const styles = {
     },
 };
 
-// react-dom.development.js:3337 The specified value "2018-10-17T00:00:00.000" does not conform to the required format, "yyyy-MM-dd".
-
 // DatePicker not yet supported in Material-UI: https://github.com/mui-org/material-ui/issues/4787
 // Fallback on browser native
 const DatePicker = ({ label, value, onChange, classes, style }) => (
     <TextField
         type="date"
         label={label}
-        defaultValue={formatTime(new Date(value))}
+        defaultValue={formatDate(value)}
         onChange={event => onChange(event.target.value)}
         style={style}
         classes={classes}
