+++ src/loaders/eventLoader.js
@@ -14,7 +14,7 @@ import {
     getBounds,
 } from '../util/geojson';
 import { EVENT_COLOR, EVENT_RADIUS } from '../constants/layers';
-import { formatTime } from '../util/helpers';
+import { formatDate } from '../util/time';
 import { cssColor } from '../util/colors';
 
 // Server clustering if more than 2000 events
@@ -50,7 +50,7 @@ const eventLoader = async layerConfig => {
         title: config.name,
         period: period
             ? getPeriodNameFromId(period.id)
-            : `${formatTime(startDate)} - ${formatTime(endDate)}`,
+            : `${formatDate(startDate)} - ${formatDate(endDate)}`,
         items: [],
     };
